moved offers' types from mocks to types
