+++ src/mocks/index.ts
@@ -1,27 +1,10 @@
 import { comments } from './comments.ts';
 import { detailOffer } from './detail.ts';
 import { offers, getOffersInNear } from './offers.ts';
-import type {
-  CommentData,
-  OfferData,
-  OfferDetailData,
-  HostUserType,
-  CityType,
-  LocationType,
-} from './types.ts';
 
 export {
   comments,
   detailOffer,
   offers,
   getOffersInNear,
-};
-
-export type {
-  CommentData,
-  OfferData,
-  OfferDetailData,
-  HostUserType,
-  CityType,
-  LocationType,
 };
