+++ src/utils/sort-offers.ts
@@ -1,4 +1,4 @@
-import { OfferData } from '../mocks';
+import { OfferData } from '../types';
 import { SORT_TYPES, SortTypesType } from '../constants';
 
 export function sortOffers(sortType: SortTypesType, offers: OfferData[]) {
